+++ CHANGELOG.md
@@ -4,9 +4,9 @@ The version numbers are according to [Semantic Versioning](http://semver.org/).
 
 ## Release v0.0.x ()
 ### Added
-
+- Added a `max_size` argument for `GoogleDocumentOcrCheckboxDetector`
 ### Fixed
-
+- Fixed a bug with checkbox detection if the image is too big by setting a default of `2048` for `max_size` of `GoogleDocumentOcrCheckboxDetector`
 ### Changed
 
 ### Removed